+++ store/db/sqlite/user_setting.go
@@ -109,16 +109,7 @@ func (d *DB) ListUserSettings(ctx context.Context, find *store.FindUserSetting) 
 }
 
 func vacuumUserSetting(ctx context.Context, tx *sql.Tx) error {
-	stmt := `
-	DELETE FROM 
-		user_setting 
-	WHERE 
-		user_id NOT IN (
-			SELECT 
-				id 
-			FROM 
-				user
-		)`
+	stmt := `DELETE FROM user_setting WHERE user_id NOT IN (SELECT id FROM user)`
 	_, err := tx.ExecContext(ctx, stmt)
 	if err != nil {
 		return err
